+++ knightbus-azureservicebus/src/KnightBus.Azure.ServiceBus/ServiceBusMessageTooLargeException.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace KnightBus.Azure.ServiceBus
+{
+    public sealed class ServiceBusMessageTooLargeException : Exception
+    {
+    }
+}